+++ changelog.md
@@ -9,6 +9,10 @@
 ## Version 0.1
 
 ### New components
+* 3 models - QL
+* Accessories.php, Categories.php, Presets.php
+* 3 controllers - QL
+* AccessoriesCtr.php, CategoriesCtr.php, PresetsCtr.php
 * csv folder - QL
 * 3 csv files - QL
 * Accessories.csv, Categories.csv, Presets.csv - QL